requirements.txt
pandas
numpy
faiss-cpu
sentence-transformers
transformers
accelerate
bitsandbytes
scikit-learn
torch
huggingface-hub
